+++ src/printer/printer.h
@@ -113,6 +113,7 @@ class CVC5_EXPORT Printer
   /** Print declare-fun command */
   virtual void toStreamCmdDeclareFunction(std::ostream& out,
                                           const std::string& id,
+                                          const std::vector<TypeNode>& argTypes,
                                           TypeNode type) const;
   /** Variant of above for a pre-existing variable */
   void toStreamCmdDeclareFunction(std::ostream& out, const Node& v) const;
@@ -122,14 +123,19 @@ class CVC5_EXPORT Printer
                                       TypeNode type,
                                       const std::vector<Node>& initValue) const;
   /** Print declare-oracle-fun command */
-  virtual void toStreamCmdDeclareOracleFun(std::ostream& out,
-                                           const std::string& id,
-                                           TypeNode type,
-                                           const std::string& binName) const;
+  virtual void toStreamCmdDeclareOracleFun(
+      std::ostream& out,
+      const std::string& id,
+      const std::vector<TypeNode>& argTypes,
+      TypeNode type,
+      const std::string& binName) const;
 
   /** Print declare-sort command */
   virtual void toStreamCmdDeclareType(std::ostream& out,
-                                      TypeNode type) const;
+                                      const std::string& id,
+                                      size_t arity) const;
+  /** Variant of above that takes the type */
+  void toStreamCmdDeclareType(std::ostream& out, TypeNode type) const;
 
   /** Print define-sort command */
   virtual void toStreamCmdDefineType(std::ostream& out,
@@ -174,7 +180,7 @@ class CVC5_EXPORT Printer
 
   /** Print declare-var command */
   virtual void toStreamCmdDeclareVar(std::ostream& out,
-                                     Node var,
+                                     const std::string& id,
                                      TypeNode type) const;
 
   /** Print synth-fun command */