+++ test/unit/main/interactive_shell_black.cpp
@@ -14,6 +14,7 @@
  */
 
 #include <cvc5/cvc5.h>
+#include <cvc5/cvc5_parser.h>
 
 #include <memory>
 #include <sstream>
@@ -24,8 +25,6 @@
 #include "options/base_options.h"
 #include "options/language.h"
 #include "options/options.h"
-#include "parser/api/cpp/command.h"
-#include "parser/api/cpp/symbol_manager.h"
 #include "test.h"
 
 using namespace cvc5::parser;